Small refacto for creating easily new integration tests
