+++ tests/integration/test_anglesandinternodes.py
@@ -1,24 +1,11 @@
 import unittest
 from pathlib import Path
 import os
-import subprocess
 import json
 import glob
-import time
 import requests
 
-def run_task(config, task, data):
-    # Check if PYOPENCL_CTX is set
-    if os.getenv('PYOPENCL_CTX') == None:
-       os.environ["PYOPENCL_CTX"] = '0'
-
-    command = ["romi_run_task", "--config", config, task, data]
-    process = subprocess.run(command, universal_newlines=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    print(process.stdout)
-    print(process.stderr)
-
-    return process
-
+from utilities import run_task
 
 class TestGeomAnglesAndInternodes(unittest.TestCase):
     def test_real_plant(self):
